requirements.txt
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
exceptiongroup==1.2.0
filelock==3.13.1
Flask==3.0.1
Flask-Cors==4.0.0
fsspec==2024.2.0
gevent==23.9.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
openai==1.13.3
packaging==23.2
pinecone-client==3.1.0
pydantic==2.6.3
pydantic_core==2.16.3
pymongo==4.6.1
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rpds-py==0.18.0
safetensors==0.4.2
sniffio==1.3.1
sympy==1.12
tokenizers==0.15.2
tqdm==4.66.2
transformers==4.38.2
triton==2.2.0
typing_extensions==4.10.0
urllib3==2.2.1
watchdog==3.0.0
Werkzeug==3.0.1
zope.event==5.0
zope.interface==6.1
